Fix loop condition, remove Write.
In some clippy changes the loop condition got flipped.

I remove the whole SeekableCompress struct: all it was is a wrapper
around Write. But this wasn't that useful and had multiple issues. To
compress data, the user should use the stream interface. This interface
also ensures that the compression is properly finalised at the end which
is a marked advantage over the Write implementation we had.

The file roundtrip_file example has been updated to demonstrate using
streams. `roundtrip_s3` is removed: we already have
`roundtrip_stream_s3` which used the streaming API. Non-streaming
approach is the same except it just buffers all the data in memory
first so it wasn't useful to have.